--- jest.config.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-	preset: 'jest-environment-obsidian',
-}
