+++ package.json
@@ -8,7 +8,7 @@
         "build": "vite build --mode production",
         "check": "svelte-check --tsconfig ./tsconfig.json",
         "wasm": "wasm-pack build ./recipe-rs --target web",
-    "test": "jest"
+        "test": "vitest"
     },
     "keywords": [],
     "author": "",
@@ -24,6 +24,7 @@
         "autoprefixer": "^10.4.20",
         "builtin-modules": "^4.0.0",
         "bun-types": "^1.1.33",
+        "jest": "^29.7.0",
         "obsidian": "^1.7.2",
         "parse-ingredient": "^1.2.1",
         "postcss": "^8.4.47",
@@ -32,7 +33,8 @@
         "tailwindcss": "^3.4.14",
         "tslib": "^2.8.0",
         "typescript": "^5.6.3",
-        "vite": "^5.4.10"
+        "vite": "^5.4.10",
+        "vitest": "^2.1.4"
     },
     "dependencies": {
         "@unocss/extractor-svelte": "^0.63.6",
